+++ src/main/java/com/example/railwaystation/classes/Game/GameLevel.java
@@ -2,7 +2,7 @@ package com.example.railwaystation.classes.Game;
 
 import com.example.railwaystation.refactored_classes.Models.CashRegister;
 import com.example.railwaystation.refactored_classes.Models.Door;
-import com.example.railwaystation.refactored_classes.Models.UserFiles.User;
+import com.example.railwaystation.classes.Moduls.Users.User;
 import javafx.scene.image.Image;
 
 import java.util.ArrayList;
