+++ src/main/java/com/example/railwaystation/classes/Helpers/QueueManager.java
@@ -3,7 +3,7 @@ package com.example.railwaystation.classes.Helpers;
 import com.example.railwaystation.classes.Game.GameLevel;
 import com.example.railwaystation.classes.Game.QueuePoligon;
 import com.example.railwaystation.refactored_classes.Models.UserFiles.OurQueue;
-import com.example.railwaystation.refactored_classes.Models.UserFiles.User;
+import com.example.railwaystation.classes.Moduls.Users.User;
 
 import java.util.Comparator;
 import java.util.List;
@@ -42,22 +42,6 @@ public class QueueManager {
         if(poligon == null)
             return;
         moveAllUsers(poligon);
-//        int index = 0;
-//        var users = poligon.get_queue()
-//                .getUsers()
-//                .stream()
-//                .toList();
-//
-//        for (var cell: poligon.get_queueCells()) {
-//            var position = cell.getPosition();
-//            if(index < users.size()) {
-//                var userPosition = users.get(index).getPosition();
-//                userPosition.setXY(position.getX(), position.getY());
-//            }
-//            else
-//                return;
-//            ++index;
-//        }
     }
 
     private void moveAllUsers(QueuePoligon poligon) {
@@ -80,33 +64,7 @@ public class QueueManager {
     }
 
     public void moveNewUser(User newUser, QueuePoligon poligon) {
-        //this.poligon = poligon;
         moveAllUsers(poligon);
-
-//        var newUserCellIndex = poligon.get_queue().getUsers().size() - 1;
-//        if(newUserCellIndex > poligon.get_queueCells().size() - 1)
-//            newUserCellIndex = poligon.get_queueCells().size() - 1;
-//
-//        var cellPosition = poligon.get_queueCells()
-//                .stream()
-//                .toList()
-//                .get(newUserCellIndex)
-//                .getPosition();
-//
-//        newUser.getPosition().setXY(
-//                cellPosition.getX(),
-//                cellPosition.getY()
-//        );
-//        var users = poligon.get_queue().getUsers().stream().toList();
-//        var cells = poligon.get_queueCells().stream().toList();
-//
-//        for (int i = 0; i < poligon.get_queueCells().size(); i++) {
-//            if(i >= users.size())
-//                return;
-//            users.get(i).getPosition().setXY(cells.get(i).getPosition().getX(),cells.get(i).getPosition().getY());
-//        }
-//        users.get(users.size()-1).getPosition().setXY(cells.get(cells.size()-1).getPosition().getX(),
-//                cells.get(cells.size()-1).getPosition().getY());
     }
 
     private boolean isPolygonFull() {
@@ -134,7 +92,6 @@ public class QueueManager {
                 .stream()
                 .sorted(Comparator.comparingInt(QueuePoligon::get_potential_count)).toList();
 
-        // if all queues are the same
         if (res.stream().allMatch(p -> p.get_potential_count() == res.get(0).get_potential_count())) {
             return res
                     .stream()
