+++ src/main/java/com/example/railwaystation/classes/Helpers/Star/DoorPolygonResolver.java
@@ -1,7 +1,7 @@
 package com.example.railwaystation.classes.Helpers.Star;
 
 import com.example.railwaystation.classes.Game.GameLevel;
-import com.example.railwaystation.classes.Helpers.Coordinates;
+import com.example.railwaystation.refactored_classes.Helpers.Coordinates;
 import com.example.railwaystation.refactored_classes.Models.Door;
 
 import java.util.HashMap;
