+++ src/main/java/com/example/railwaystation/classes/Helpers/Star/StarAlgorithm.java
@@ -1,7 +1,6 @@
 package com.example.railwaystation.classes.Helpers.Star;
 
-import com.example.railwaystation.classes.Game.CellState;
-import com.example.railwaystation.classes.Helpers.Coordinates;
+import com.example.railwaystation.refactored_classes.Helpers.Coordinates;
 
 import java.util.*;
