+++ src/main/java/com/example/railwaystation/classes/Logic/Game.java
@@ -3,15 +3,15 @@ package com.example.railwaystation.classes.Logic;
 import com.example.railwaystation.classes.Game.AssetsReader;
 import com.example.railwaystation.classes.Game.GameLevel;
 import com.example.railwaystation.classes.Game.LevelReader;
-import com.example.railwaystation.classes.Helpers.Coordinates;
+import com.example.railwaystation.refactored_classes.Helpers.Coordinates;
 import com.example.railwaystation.classes.Helpers.Star.Node;
 import com.example.railwaystation.refactored_classes.Interfaces.Generator;
 import com.example.railwaystation.refactored_classes.Models.Door;
 import com.example.railwaystation.refactored_classes.Models.UserFiles.OurQueue;
-import com.example.railwaystation.classes.Rendering.ResourceManagerUser;
+import com.example.railwaystation.refactored_classes.Rendering.ResourceManagerUser;
 import com.example.railwaystation.refactored_classes.Models.UserFiles.Priority;
 import com.example.railwaystation.refactored_classes.Models.UserFiles.PrototypeRegistry;
-import com.example.railwaystation.refactored_classes.Models.UserFiles.User;
+import com.example.railwaystation.classes.Moduls.Users.User;
 import com.example.railwaystation.refactored_classes.Models.UserFiles.UserType;
 
 import java.util.Collection;
