+++ src/main/java/com/example/railwaystation/refactored_classes/Interfaces/Generator.java
@@ -1,7 +1,7 @@
 package com.example.railwaystation.refactored_classes.Interfaces;
 
 
-import com.example.railwaystation.refactored_classes.Models.UserFiles.User;
+import com.example.railwaystation.classes.Moduls.Users.User;
 
 public interface Generator {
     public User generateUser();
