+++ src/main/java/com/example/railwaystation/refactored_classes/Models/CashRegister.java
@@ -1,9 +1,9 @@
 package com.example.railwaystation.refactored_classes.Models;
 
-import com.example.railwaystation.classes.Helpers.Coordinates;
+import com.example.railwaystation.refactored_classes.Helpers.Coordinates;
 import com.example.railwaystation.refactored_classes.Models.UserFiles.OurQueue;
 import com.example.railwaystation.refactored_classes.MovingFiles.StopWatch;
-import com.example.railwaystation.refactored_classes.Models.UserFiles.User;
+import com.example.railwaystation.classes.Moduls.Users.User;
 import javafx.scene.image.Image;
 
 public class CashRegister extends GameObject {
