+++ src/main/java/com/example/railwaystation/refactored_classes/Rendering/CanvasRendering.java
@@ -1,18 +1,16 @@
-package com.example.railwaystation.classes.Rendering;
+package com.example.railwaystation.refactored_classes.Rendering;
 
 import com.example.railwaystation.refactored_classes.UI.CanvasManager.Camera2D;
-import com.example.railwaystation.classes.Helpers.Coordinates;
+import com.example.railwaystation.refactored_classes.Helpers.Coordinates;
 import com.example.railwaystation.classes.Logic.Game;
 import javafx.scene.canvas.Canvas;
 import javafx.scene.image.Image;
+import javafx.scene.paint.Color;
 
 
 public class CanvasRendering implements Rendering {
     private Canvas _canvas;
     private Camera2D _camera;
-    public Camera2D get_camera() {
-        return _camera;
-    }
 
     public void set_camera(Camera2D _camera) {
         this._camera = _camera;
@@ -37,7 +35,7 @@ public class CanvasRendering implements Rendering {
     @Override
     public void DrawGrid(double width, double height, double step_x, double step_y) {
         var ctx = _canvas.getGraphicsContext2D();
-//        ctx.setStroke(Color.BLUEVIOLET);
+        ctx.setStroke(Color.BLUEVIOLET);
         double width_low = 0 - _camera.get_position().getX();
         double width_high = width*_camera.get_zoom() - _camera.get_position().getX();
 
@@ -47,12 +45,12 @@ public class CanvasRendering implements Rendering {
         step_x *=_camera.get_zoom();
         step_y *=_camera.get_zoom();
 
-//        for(double x=width_low; x <= width_high; x+=step_x){
-//            _canvas.getGraphicsContext2D().strokeLine(x,height_low,x,height_high);
-//        }
-//        for (double y=height_low; y <= height_high; y+=step_y){
-//            _canvas.getGraphicsContext2D().strokeLine(width_low,y,width_high,y);
-//        }
+        for(double x=width_low; x <= width_high; x+=step_x){
+            _canvas.getGraphicsContext2D().strokeLine(x,height_low,x,height_high);
+        }
+        for (double y=height_low; y <= height_high; y+=step_y){
+            _canvas.getGraphicsContext2D().strokeLine(width_low,y,width_high,y);
+        }
     }
