+++ src/main/java/com/example/railwaystation/refactored_classes/Rendering/ResourceManagerCashRegister.java
@@ -1,10 +1,11 @@
-package com.example.railwaystation.classes.Rendering;
+package com.example.railwaystation.refactored_classes.Rendering;
 
 import javafx.scene.image.Image;
+
 import java.util.HashMap;
 import java.util.Map;
 
-public class ResourceManagerQueueCell {
+public class ResourceManagerCashRegister {
     private static final Map<String, Image> sprites = new HashMap<>();
     public static Image getSprite(String key){
         return sprites.get(key);
