+++ src/main/java/com/example/railwaystation/refactored_classes/Rendering/ResourceManagerDoor.java
@@ -1,11 +1,12 @@
-package com.example.railwaystation.classes.Rendering;
+package com.example.railwaystation.refactored_classes.Rendering;
 
 import javafx.scene.image.Image;
+
 import java.util.HashMap;
 import java.util.Map;
 
-public class ResourceManagerUser {
-    private static Map<String, Image> sprites = new HashMap<>();
+public class ResourceManagerDoor {
+    private static final Map<String, Image> sprites = new HashMap<>();
     public static Image getSprite(String key){
         return sprites.get(key);
     }
